refactor(checkout menu): make menu work with simple implementation
